better connection handling

+++ src/mcp/constants.rs
@@ -1,4 +1,5 @@
-pub(super) const HANDSHAKE: u8 = 0x00;
+pub(super) const HANDSHAKE: u32 = 0x00;
+
 pub const VARINT_SEGMENT_BITS: u8 = 0x7F;
 pub const VARINT_SEGMENT_BITS_I32: i32 = 0x7F;
 pub const VARINT_CONTINUE_BIT: u8 = 0x80;
